Move mount validation into mount module
