+++ northstar-runtime/src/npk/manifest/validation.rs
@@ -2,17 +2,10 @@ use crate::{
     common::non_nul_string::NonNulString,
     seccomp::{Seccomp, SyscallRule},
 };
-use itertools::Itertools;
-use std::{
-    collections::HashMap,
-    path::{Component, Component::RootDir, Path},
-};
+use std::collections::HashMap;
 use validator::ValidationError;
 
-use super::{
-    mount::{Mount, MountOption, MountPoint},
-    Manifest,
-};
+use super::Manifest;
 
 /// Maximum number of environment variables
 const MAX_ENV_VARS: usize = 64;
@@ -76,45 +69,6 @@ pub fn env(env: &HashMap<NonNulString, NonNulString>) -> Result<(), ValidationEr
     } else {
         Ok(())
     }
-}
-
-pub fn mounts(mounts: &HashMap<MountPoint, Mount>) -> Result<(), ValidationError> {
-    // Check for relative and overlapping bind mounts
-    let mut prev_comps = vec![RootDir];
-    mounts
-        .iter()
-        .filter(|(_, m)| matches!(m, Mount::Bind(_)))
-        .map(|(p, _)| p)
-        .sorted()
-        .try_for_each(|p| {
-            let p: &Path = p.as_ref();
-            if p.is_relative() {
-                return Err(ValidationError::new("mount points must not be relative"));
-            }
-            // Check for overlapping bind mount paths by checking if one path is the prefix of the next one
-            let curr_comps: Vec<Component> = p.components().collect();
-            let prev_too_short = prev_comps.len() <= 1; // Two mount paths both starting with '/' is not considered an overlap
-            let prev_too_long = prev_comps.len() > curr_comps.len(); // A longer path cannot be the prefix of a shorter one
-
-            if !prev_too_short && !prev_too_long && prev_comps == curr_comps[..prev_comps.len()] {
-                Err(ValidationError::new("mount points must not overlap"))
-            } else {
-                prev_comps = curr_comps;
-                Ok(())
-            }
-        })?;
-
-    // Check for recursive non bind mounts
-    mounts.iter().map(|(_, m)| m).try_for_each(|m| match m {
-        // Recursive bind mounts are allowed but not resources
-        Mount::Resource(m) if m.options.contains(&MountOption::Rec) => Err(ValidationError::new(
-            "non bind mounts must not be recursive",
-        )),
-        Mount::Resource(m) if !m.dir.starts_with('/') => Err(ValidationError::new(
-            "resource directory options must be absolute",
-        )),
-        _ => Ok(()),
-    })
 }
 
 /// Validate seccomp rules
